+++ crates/python/src/lib.rs
@@ -31,8 +31,6 @@ create_init_submodule! {
         execution_data::PyRegisterMatrixConversionError
     ],
     funcs: [
-        api::py_submit,
-        api::py_submit_async,
         api::py_retrieve_results,
         api::py_retrieve_results_async,
         api::py_list_quantum_processors,
