+++ crates/python/src/qpu/mod.rs
@@ -7,6 +7,7 @@ pub mod isa;
 pub mod quilc;
 mod result_data;
 pub mod rewrite_arithmetic;
+pub mod runner;
 pub mod translation;
 
 use isa::QCSISAError;
@@ -19,6 +20,7 @@ create_init_submodule! {
         "isa": isa::init_submodule,
         "quilc": quilc::init_submodule,
         "rewrite_arithmetic": rewrite_arithmetic::init_submodule,
+        "runner": runner::init_submodule,
         "translation": translation::init_submodule
     ],
 }
